rework ui components
