+++ app/src/main/java/com/f4/mypet/ui/screens/procedure/ProcedureScreen.kt
@@ -38,7 +38,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.f4.mypet.R
 import com.f4.mypet.navigation.Routes
-import com.f4.mypet.ui.MyPetTopBar
+import com.f4.mypet.ui.components.MyPetTopBar
 
 
 @Composable
