+++ app/src/main/java/com/f4/mypet/ui/screens/profile/ListProfileScreen.kt
@@ -49,8 +49,8 @@ import androidx.navigation.NavHostController
 import com.f4.mypet.R
 import com.f4.mypet.navigation.BottomBarRoutes
 import com.f4.mypet.navigation.Routes
-import com.f4.mypet.ui.CustomSnackBar
-import com.f4.mypet.ui.MyPetTopBar
+import com.f4.mypet.ui.components.MyPetSnackBar
+import com.f4.mypet.ui.components.MyPetTopBar
 import com.f4.mypet.ui.theme.BlueCheckbox
 import com.f4.mypet.ui.theme.GreenButton
 import com.f4.mypet.ui.theme.LightBlueBackground
@@ -80,7 +80,7 @@ fun ListProfileScreen(
             SnackbarHost(
                 hostState = snackbarHostState
             ) {
-                CustomSnackBar(it.visuals.message)
+                MyPetSnackBar(it.visuals.message)
             }
         }
     ) { innerPadding ->
